\name{bupa}
\alias{bupa}
\docType{data}
\title{
Bupa Dataset
}
\description{
Bupa Dataset
}
\usage{data(bupa)}
\format{
  A data frame with 345 observations on the following 14 variables.
  \describe{
    \item{\code{Case}}{a numeric vector}
    \item{\code{V1}}{a numeric vector}
    \item{\code{V2}}{a numeric vector}
    \item{\code{V3}}{a numeric vector}
    \item{\code{V4}}{a numeric vector}
    \item{\code{V5}}{a numeric vector}
    \item{\code{V6}}{a numeric vector}
    \item{\code{y}}{a numeric vector}
    \item{\code{V1.2}}{a numeric vector}
    \item{\code{V2.2}}{a numeric vector}
    \item{\code{V3.2}}{a numeric vector}
    \item{\code{V4.2}}{a numeric vector}
    \item{\code{V5.2}}{a numeric vector}
    \item{\code{V6.2}}{a numeric vector}
  }
}
\examples{
data(bupa)
## maybe str(bupa) ; plot(bupa) ...
}
\keyword{datasets}
