\name{cleveland}
\alias{cleveland}
\docType{data}
\title{
Cleveland Dataset
}
\description{
Cleveland Dataset
}
\usage{data(cleveland)}
\format{
  A data frame with 297 observations on the following 15 variables.
  \describe{
    \item{\code{V1}}{a numeric vector}
    \item{\code{V2}}{a numeric vector}
    \item{\code{V3}}{a numeric vector}
    \item{\code{V4}}{a numeric vector}
    \item{\code{V5}}{a numeric vector}
    \item{\code{V6}}{a numeric vector}
    \item{\code{V7}}{a numeric vector}
    \item{\code{V8}}{a numeric vector}
    \item{\code{V9}}{a numeric vector}
    \item{\code{V10}}{a numeric vector}
    \item{\code{V11}}{a numeric vector}
    \item{\code{V12}}{a numeric vector}
    \item{\code{V13}}{a numeric vector}
    \item{\code{case}}{a logical vector}
    \item{\code{y}}{a numeric vector}
  }
}
\examples{
data(cleveland)
## maybe str(cleveland) ; plot(cleveland) ...
}
\keyword{datasets}
