\name{kyphosis}
\alias{kyphosis}
\docType{data}
\title{
Kyphosis Data
}
\description{
Kyphosis Data
}
\usage{data(kyphosis)}
\format{
  A data frame with 81 observations on the following 5 variables.
  \describe{
    \item{\code{Kyphosis}}{a factor with levels \code{absent} \code{present}}
    \item{\code{Age}}{a numeric vector}
    \item{\code{Number}}{a numeric vector}
    \item{\code{Start}}{a numeric vector}
    \item{\code{y}}{a numeric vector}
  }
}
\examples{
data(kyphosis)
## maybe str(kyphosis) ; plot(kyphosis) ...
}
\keyword{datasets}
